+++ examples/remote_spawn/table.c
@@ -109,7 +109,7 @@ void table_receive(struct bsal_actor *actor, struct bsal_message *message)
 
         if (table1->done == bsal_vector_size(&table1->spawners)) {
             printf("actor %d kills %d to %d\n",
-                           name, 0, bsal_vector_size(&table1->spawners) - 1);
+                           name, 0, (int)bsal_vector_size(&table1->spawners) - 1);
             bsal_message_init(message, TABLE_DIE, 0, NULL);
             bsal_actor_send_range_standard(actor, &table1->spawners, message);
         }
